+++ CHANGES.rst
@@ -5,6 +5,7 @@ Kahn changelog
 
 in progress
 ===========
+- Fix UDP message encoding
 
 
 2022-07-25 0.0.0